+++ client/experiment-dispatcher/index.tsx
@@ -8,20 +8,8 @@ import { setCurrentUser } from 'calypso/state/current-user/actions';
 import { getInitialState, getStateFromCache } from 'calypso/state/initial-state';
 import initialReducer from 'calypso/state/reducer';
 import { setStore } from 'calypso/state/redux-store';
-import { experimentsCatalog } from './experiments';
 import 'calypso/components/environment-badge/style.scss';
-
-/**
- * Gets the experiment manifest by determining its slug from the URL then retrieving it from the catalog.
- */
-const getManifestFromUrl = () => {
-	const slug = window.location.pathname.split( '/' )[ 2 ] as keyof typeof experimentsCatalog;
-	const experiment = slug in experimentsCatalog && experimentsCatalog[ slug ];
-	if ( ! experiment ) {
-		throw new Error( `Experiment with slug ${ slug } not found` );
-	}
-	return experiment;
-};
+import { redirectWithParams, getManifestFromUrl } from './utils';
 
 window.AppBoot = async () => {
 	/**
@@ -45,7 +33,7 @@ window.AppBoot = async () => {
 
 		// Redirect to control for null assignment.
 		if ( ! assignment.variationName ) {
-			window.location.replace( experimentManifest.variants[ 'control' ].url );
+			redirectWithParams( experimentManifest.variants[ 'control' ].url );
 			return;
 		}
 
@@ -53,14 +41,14 @@ window.AppBoot = async () => {
 			alert(
 				`Variation ${ assignment.variationName } not found in experiment ${ experimentManifest.title }`
 			);
-			window.location.replace( experimentManifest.variants[ 'control' ].url );
+			redirectWithParams( experimentManifest.variants[ 'control' ].url );
 			return;
 		}
 
 		// Redirect to treatment.
-		window.location.replace( experimentManifest.variants[ assignment.variationName ].url );
+		redirectWithParams( experimentManifest.variants[ assignment.variationName ].url );
 	} catch ( e ) {
 		// In case anything goes wrong, don't leave users stranded and go to control.
-		window.location.replace( experimentManifest.variants[ 'control' ].url );
+		redirectWithParams( experimentManifest.variants[ 'control' ].url );
 	}
 };
